refactor: changed file structure to match the one documented
